fishion default prompt units and the order can be selected
Users can further customized the order of units or their existense
in the prompt by using fishion_prompt_unit variable.

Any word used in this variable will be used to construct the default
prompt, in order. unit names are replaced properly with the unit.
Available units are:
 - user: current user name
 - host: the hostname
 - cwd: current working directory (or pwd)
 - status: the exit code of last running command

By default the promopt would be printed as
user@host cwd[status]

and prefix and suffix are always there.