Makefile install/uninstall supports manaing the prompt
So it's easier to install or remove the fishion prompt
by running make install (or uninstall).

To prevent modifying the user prompt directly, the Makefile
will ask the user to do the final step.

Also added a name/author comment to the fish_prompt file because
that's used by fish web config tool to show the promopt in the list.